Create License file
I love your project and I'm sure many others do as well. I would really like to reuse it in a student project, but I need a license for that. I've suggested the MIT license. If you don't mind, please include this in your project or provide another license. This would be an immense help! Thanks :)